fixed basket

+++ app/src/main/java/com/example/burgerapplication/repository/CartRepository.kt
@@ -4,6 +4,7 @@ import android.util.Log
 import com.example.burgerapplication.api.ProductApiService
 import com.example.burgerapplication.dao.CartDao
 import com.example.burgerapplication.dto.Cart
+import com.example.burgerapplication.dto.CartResponse
 import com.example.burgerapplication.dto.Product
 import com.example.burgerapplication.entity.CartEntity
 import javax.inject.Inject
@@ -12,18 +13,28 @@ class CartRepository@Inject constructor(
     private val productApiService: ProductApiService,
     private val cartDao: CartDao
 ) {
-    suspend fun sendCartToServer(cartItems: List<CartEntity>, token: String) {
+    suspend fun sendCartToServer(cartItems: List<CartEntity>, token: String): CartResponse? {
         val cartData = cartItems.map { Cart(it.productId, it.quantity) }
 
-        val response = productApiService.sendCart(token, cartData)
-        if (response.isSuccessful) {
-            val sentCart = response.body()
+        val response = productApiService.sendCart("Bearer $token", cartData)
+        return if (response.isSuccessful) {
+            val cartResponse = response.body()
+            if (cartResponse != null) {
+                Log.d("CartRepository", "Final Price: ${cartResponse.finalPrice}")
+                Log.d("CartRepository", "Points: ${cartResponse.points}")
+                Log.d("CartRepository", "Products: ${cartResponse.products}")
+            }
+            cartResponse
         } else {
             Log.e("CartRepository", "Error sending cart: ${response.errorBody()?.string()}")
+            null
         }
     }
+
     suspend fun saveCartLocally(product: Product, quantity: Int) {
-        val cartEntity = CartEntity(productId = product.id, quantity = quantity)
+        val existingQuantity = cartDao.getQuantityByProductId(product.id) ?: 0
+        val newQuantity = existingQuantity + quantity
+        val cartEntity = CartEntity(productId = product.id, quantity = newQuantity)
         cartDao.insert(cartEntity)
     }
 
@@ -36,14 +47,17 @@ class CartRepository@Inject constructor(
     }
 
     suspend fun removeFromCartLocal(product: Product) {
-        cartDao.deleteByProductId(product.id)
+        val currentQuantity = cartDao.getQuantityByProductId(product.id) ?: 0
+        if (currentQuantity > 1) {
+            val newQuantity = currentQuantity - 1
+            cartDao.insert(CartEntity(productId = product.id, quantity = newQuantity))
+        } else {
+            cartDao.deleteByProductId(product.id)
+        }
     }
 
     suspend fun getAllCartItems(): List<CartEntity> {
         return cartDao.getAllCartItems()
     }
 
-//    suspend fun clearLocalCart() {
-//        cartDao.clearCart()
-//    }
 }
